+++ scripts.config.ts
@@ -19,18 +19,18 @@ const SENetworkNames = SEMessengerSpec.ipfsData.serviceTicker.values;
 export const scripts: ScriptsConfiguration = {
   deploy_sla: [
     {
-      sloValue: 7,
+      sloValue: 100,
       sloType: SLO_TYPE.GreaterThan,
       whitelisted: false,
       periodType: PERIOD_TYPE.WEEKLY,
       messengerContract: CONTRACT_NAMES.SEMessenger,
       initialPeriodId: 0,
-      finalPeriodId: 9,
+      finalPeriodId: 3,
       extraData: [],
       initialTokenSupply: '10000000',
       initialTokenSupplyDivisor: 100,
       deployerStakeTimes: 100,
-      notDeployerStakeTimes: 2,
+      notDeployerStakeTimes: 1,
       leverage: 50,
       serviceMetadata: {
         serviceName: 'P-OPS',
@@ -115,7 +115,7 @@ export const scripts: ScriptsConfiguration = {
       },
     },
     {
-      sloValue: 11,
+      sloValue: 200,
       sloType: SLO_TYPE.GreaterThan,
       whitelisted: false,
       periodType: PERIOD_TYPE.WEEKLY,